+++ NEWS.md
@@ -15,6 +15,10 @@
   this automatically. It also no longer generates an package-level doc
   template, instead call `use_package_doc()`.
 
+* New `use_data()` makes it easy to include data in a package, either 
+  in `data/` (for exported datasets) or in `R/sysdata.rda` (for internal
+  data). (#542)
+
 * New `use_rcpp()` sets up a package to use Rcpp.
 
 * New `use_knitr()` sets up a package to use knitr for vignettes.